Fix android-studio zap :delete path
Commit cd42121c21b56a1563979a65ba4e3fddcae5a2e5 implemented the zap
stanza, deriving it from the following Stack Overflow answer:
http://stackoverflow.com/a/18458893

Since the commands in the SO answer did not have the paths in quotes,
the space in ~/Library/Application\ Support/AndroidStudio* was escaped
with a backslash. The backslash was imported in the commit as well, even
though the path was in quotes, which causes the backslash to be
interpreted literally. This commit fixes that by removing the backslash.